ProjectTemplate {ProjectTemplate}	R Documentation

Automates the creation of new statistical analysis projects.

Description

ProjectTemplate provides functions to automatically build a directory structure for a new R project. Using this structure, ProjectTemplate automates data loading, preprocessing, library importing and unit testing.

Details

A list of file types is supported by default. See the preinstalled.readers for more information which those are.

References

This code is inspired by the skeleton structure used by Ruby on Rails.

Examples

library('ProjectTemplate')

## Not run: create.project('project_name')

setwd('project_name')
load.project()
## End(Not run)

[Package ProjectTemplate version 0.9.3 Index]