arff.reader {ProjectTemplate}	R Documentation

Read the Weka file format.

Description

This function will load a data set stored in the Weka file format into the specified global variable binding.

Usage

arff.reader(data.file, filename, variable.name)

Arguments

data.file
The name of the data file to be read.

filename
The path to the data set to be loaded.

variable.name
The name to be assigned to in the global environment.

Value

No value is returned; this function is called for its side effects.

Examples

## Not run: arff.reader('example.arff', 'data/example.arff', 'example')

[Package ProjectTemplate version 0.8.2 Index]