cache.name {ProjectTemplate}	R Documentation

Translate a variable name into a file name for caching.

Description

This function will translate a variable name into a form that is suitable as a filename on most OS's.

Usage

cache.name(data.filename)

Arguments

data.filename
The variable name to be translated into a filename.

Value

A translated variable name.

Examples

library('ProjectTemplate')

## Not run: cache.name('example.1')

[Package ProjectTemplate version 0.9.3 Index]