cache.project {ProjectTemplate}	R Documentation

Cache a project's data sets in binary format.

Description

This function will cache all of the data sets that were loaded by the load.project function in a binary format that is easier to load quickly. This is particularly useful for data sets that you've modified during a slow munging process that does not need to be repeated.

Usage

cache.project()

Value

No value is returned; this function is called for its side effects.

See Also

create.project, load.project, get.project, show.project

Examples

library('ProjectTemplate')
## Not run: load.project()

cache.project()
## End(Not run)

[Package ProjectTemplate version 0.9.3 Index]