create.template {ProjectTemplate}	R Documentation

Create a new template

Description

This function writes a skeleton directory structure for creating your own custom templates.

Usage

create.template(target, source = "minimal")

Arguments

target
Name of the new template. It is created under the directory specified by options('ProjectTemplate.templatedir'), or, when missing, in the current directory.

source
Name of an existing template to copy, defaults to the built in 'minimal' template.

[Package ProjectTemplate version 0.9.3 Index]