db.reader {ProjectTemplate}	R Documentation

Read a SQlite3 database with a (.db) file extension.

Description

This function will load all of the data sets stored in the SQlite3 database into the global environment. If you want to specify a single table or query to execute against the database, move it elsewhere and use a .sql file interpreted by sql.reader.

Usage

db.reader(data.file, filename, variable.name)

Arguments

data.file
The name of the data file to be read.

filename
The path to the data set to be loaded.

variable.name
The name to be assigned to in the global environment.

Value

No value is returned; this function is called for its side effects.

Examples

library('ProjectTemplate')

## Not run: db.reader('example.db', 'data/example.db', 'example')

[Package ProjectTemplate version 0.8.2 Index]