dbf.reader {ProjectTemplate}	R Documentation

Read an XBASE file with a .dbf file extension.

Description

This function will load all of the data sets stored in the specified XBASE file into the global environment.

Usage

dbf.reader(data.file, filename, variable.name)

Arguments

data.file
The name of the data file to be read.

filename
The path to the data set to be loaded.

variable.name
The name to be assigned to in the global environment.

Value

No value is returned; this function is called for its side effects.

Examples

library('ProjectTemplate')

## Not run: dbf.reader('example.dbf', 'data/example.dbf', 'example')

[Package ProjectTemplate version 0.8.2 Index]