.attach.or.add.namespace {ProjectTemplate}	R Documentation

Attach a package or add a namespace

Description

Internal method to attach a package or only add the namespace.

Usage

.attach.or.add.namespace(package.name, attach)

Arguments

package.name
name of the package to load, as a character vector

attach
boolean indicating whether to attach the package in the global namespace

Value

Boolean indicating whether the package was successfully loaded

[Package ProjectTemplate version 0.9.3 Index]