.cache.filename {ProjectTemplate}	R Documentation

Construct the file names for the cache and hash

Description

Construct the file names for the cache and hash

Usage

.cache.filename(variable)

Arguments

variable
Variable name for which to construct file names

Details

The returned object is a list with two fields:

obj: The path to the file in which the variable contents will be saved;

hash: The path to the file in which the cache metadata will be stored.

Value

A list with file names

[Package ProjectTemplate version 0.9.3 Index]