.check.version {ProjectTemplate}	R Documentation

Compare the project version with the current ProjectTemplate version

Description

Compare the project version with the current ProjectTemplate version

Usage

.check.version(config, warn.migrate = TRUE)

Arguments

config
Project configuration

warn.migrate
Logical indicating whether a warning should be raised if the project version is older than the installed version of ProjectTemplate.

Value

0 if the numbers are equal, -1 if b is later and 1 if a is later (analogous to the C function strcmp).

[Package ProjectTemplate version 0.9.3 Index]