.create.cache.hash {ProjectTemplate}	R Documentation

Create a data.frame with the cache metadata

Description

Create a data.frame with the cache metadata

Usage

.create.cache.hash(variable, depends, CODE)

Arguments

variable
Name of the variable to be cached

depends
Vector of variable names of dependencies for the variable to be cached, optional.

CODE
Code block to generate variable, registered as a dependency, optional.

Details

The hashes for the various objects are calculated using the .cache.hash function.

Value

data.frame containing the variable name and its dependencies, with the corresponding hashes appended.

See Also

.cache.hash

[Package ProjectTemplate version 0.9.3 Index]