.create.project.existing {ProjectTemplate}	R Documentation

Create a project structure

Description

.create.project.existing creates a project directory structure inside an existing directory with the default files from a given template.

.create.project.new first creates a new directory and then passes further control to .create.project.existing. In case the project creation fails, the newly created directory is cleaned up.

Usage

.create.project.existing(
  project.name,
  merge.strategy,
  template,
  rstudio.project
)

.create.project.new(project.name, template, rstudio.project)

Arguments

project.name
Character vector with the name of the project directory

merge.strategy
Character vector determining whether the directory should be empty or is allowed to contain non-conflicting files

template
Name of the template from which the project should be created

rstudio.project
Logical indicating whether an .Rproj file should be created

Value

No value is returned; this function is called for its side effects.

See Also

create.project, create.template

[Package ProjectTemplate version 0.9.3 Index]