.evaluate.code {ProjectTemplate}	R Documentation

Run code and assign the results to variable

Description

Run code and assign the results to variable

Usage

.evaluate.code(variable, CODE)

Arguments

variable
variable name in which to store the result of CODE

CODE
code block that returns a result which can be stored in a variable

Details

No error handling is done on the executed code, nor is the

[Package ProjectTemplate version 0.9.3 Index]