.get.template {ProjectTemplate}	R Documentation

Get the location of a template from its name

Description

Checks the configured option('ProjectTemplate.templatedir') for the template. If no matching template was found the system templates are checked, and finally the current directory is checked. If no template was found with the given name an error is raised.

Usage

.get.template(template)

Arguments

template
Character vector containing the name of the template

Value

Character vector containing the location of the template. If no template was found by the given name an error is raised.

[Package ProjectTemplate version 0.9.3 Index]