.is.ProjectTemplate {ProjectTemplate}	R Documentation

Test whether a given path is a ProjectTemplate project

Description

Test whether a given path is a ProjectTemplate project

Usage

.is.ProjectTemplate(path = getwd())

Arguments

path
Directory to check, defaults to the current working directory.

Value

Logical indicating whether the given path is a valid project.

[Package ProjectTemplate version 0.9.3 Index]