.is.dir {ProjectTemplate}	R Documentation

Check if path is an existing directory

Description

Checks if a given path exists, and if so if it is a directory.

Usage

.is.dir(path)

Arguments

path
Character vector containing the path to the directory to check.

Value

Logical indicating a valid directory was passed.

[Package ProjectTemplate version 0.9.3 Index]