.list.files.and.dirs {ProjectTemplate}	R Documentation

List all files and directories, excluding .. and .

Description

Creates a directory listing of a given path, including hidden files and subdirectories, but excluding the .. and . aliases.

Usage

.list.files.and.dirs(path)

Arguments

path
Character vector indicating the path to the parent folder of which the contents should be listed.

Value

Directory listing of path

[Package ProjectTemplate version 0.9.3 Index]