.load.data {ProjectTemplate}	R Documentation

Load the data from the cache and data directories

Description

Gets the list of available variables in cache/ and data/ and loads the data in memory. Data from the cache is loaded first, then in alphabetical order.

Usage

.load.data(config, my.project.info)

Arguments

config
Named list containing the project configuration

my.project.info
Named list containing the project information

Value

Returns my.project.info amended with the new information.

[Package ProjectTemplate version 0.9.3 Index]