.load.libraries {ProjectTemplate}	R Documentation

Load the libraries listed in the configuration into memory

Description

Load the libraries listed in the libraries entry in global.dcf and add the library names to the project.info.

Usage

.load.libraries(config, my.project.info)

Arguments

config
Named list containing the project configuration

my.project.info
Named list containing the project information

Value

Returns my.project.info amended with the new information.

[Package ProjectTemplate version 0.9.3 Index]