.parse.extensions {ProjectTemplate}	R Documentation

Match readers to the extensions of the data files

Description

Match readers to the extensions of the data files

Usage

.parse.extensions(data.files, config)

Arguments

data.files
a vector of paths to data files

Value

A list of readers and varnames

[Package ProjectTemplate version 0.9.3 Index]