.prepare.data.ignore.regex {ProjectTemplate}	R Documentation

Prepare a regular expression for matching files to be ignored

Description

Constructs a single regular expression for matching file names in data that should not be imported. It can detect literal names, globs with wildcards and regular expressions.

Usage

.prepare.data.ignore.regex(ignore_files)

Arguments

ignore_files
A comma separated character vector that lists all patterns to be matched for ignoring

Value

A chained regular expression that matches all patterns in the ignore_files variable.

[Package ProjectTemplate version 0.9.3 Index]