.provide.directory {ProjectTemplate}	R Documentation

Make sure a required directory exists before usage

Description

Checks if the requested directory exists, and if not creates the directory. In the latter case a warning is raised.

Usage

.provide.directory(name)

Arguments

name
Character vector containing the name of the required directory.

Value

No value is returned; this function is called for its side effects.

[Package ProjectTemplate version 0.9.3 Index]