.read.cache.info {ProjectTemplate}	R Documentation

Read metadata for a variable in the cache

Description

Read metadata for a variable in the cache

Usage

.read.cache.info(variable)

Arguments

variable
Variable name for which to look up the metadata

Details

The returned object is a list with two fields:

in.cache: Logical indicating whether the requested variable was found in the cache

hash: A data.frame as was created by .create.cache.hash

Value

list with metadata, see Details for more info.

[Package ProjectTemplate version 0.9.3 Index]