.remove.sticky.vars {ProjectTemplate}	R Documentation

Remove variables to keep from a list of candidates for removal

Description

Remove variables to keep from a list of candidates for removal

Usage

.remove.sticky.vars(names, keep)

Arguments

names
character vector of variable names that are candidate for removal

keep
character vector of variable names that should not be removed

Details

If the sticky_variables option is part of the config variable the config variable itself is added to the list of variables to keep. Also all variables listed in config$sticky_variables in a comma separated list are added to keep.

Value

A character vector containing the variables to remove.

[Package ProjectTemplate version 0.9.3 Index]