.require.package {ProjectTemplate}	R Documentation

Require internal package

Description

Internal method to require a package that is necessary for the internal functioning of ProjectTemplate. Never attaches the package unless configured to do so in global.dcf (which throws a warning).

Usage

.require.package(package.name)

Arguments

package.name
name of the package to load, as a character vector

Value

No value is returned; this function is called for its side effects.

[Package ProjectTemplate version 0.9.3 Index]