.stopifnotproject {ProjectTemplate}	R Documentation

Raise an error if given path is not a valid project

Description

Function to stop processing if the path is not a Project Template return the project name if it is a Project Template directory.

Usage

.stopifnotproject(additional_message = "", path = getwd())

Arguments

additional_message
Optional message to show if the given path is not a valid project

path
Path to check if it is a valid project

Value

Project name if it is a valid Project.

[Package ProjectTemplate version 0.9.3 Index]