.stopifproject {ProjectTemplate}	R Documentation

Raise an error if given path is a valid project

Description

Function to stop processing if the path is a Project Template.

Usage

.stopifproject(additional_message = "", path = getwd())

Arguments

additional_message
Optional message to show if the given path is not a valid project

path
Path to check if it is a valid project

Value

No value is returned; this function is called for its side effects

[Package ProjectTemplate version 0.9.3 Index]