.var.diff.from {ProjectTemplate}	R Documentation

Compare sets of variable names

Description

Compare the variables (excluding functions) in the global env with a passed in string of names and return the set difference.

Usage

.var.diff.from(given.var.list = "", env = .TargetEnv)

Arguments

given.var.list
Character vector of variable names

env
Environment in which to compare the sets of variables

[Package ProjectTemplate version 0.9.3 Index]