.write.cache {ProjectTemplate}	R Documentation

Write a variable and its metadata to cache

Description

Write a variable and its metadata to cache

Usage

.write.cache(cache.hash, ...)

Arguments

cache.hash
a data.frame with metadata about the variable, see details for more information.

...
extra parameters passed to save.

Details

cache.hash is a data frame with two columns: variable and hash.
Row name VAR is the name of the variable to save.
Row name CODE is the hash value of the code to compute variable.
Row name DEPENDS.* are the dependent variables that CODE depends on.c
The helper function .create.cache.hash creates a suitable dataframe

Value

No value is returned, this function is called for its side effects.

[Package ProjectTemplate version 0.9.3 Index]