list.data {ProjectTemplate}	R Documentation

Listing the data for the current project

Description

This function produces a data.frame of all data files in the project, with meta data on if and how the file will be loaded by load.project.

Usage

list.data(...)

Arguments

...
Named arguments to override configuration from config/global.dcf and lib/global.R.

Details

The returned data.frame contains the following variables, with one observation per file in data/:

filename 	 Character variable containing the filename relative to data/ directory.

varname 	 Character variable containing the name of the variable into which the file will be imported. *

is_ignored 	 Logical variable that indicates whether the file. is ignored through the data_ignore option in the configuration

is_directory 	 Logical variable that indicates whether the file is a directory.

is_cached 	 Logical variable that indicates whether the file is already available in the cache/ directory.

cached_only 	 Logical variable that indicates whether the variable is only available in the cache/ directory. This occurs when calling the cache function with a code fragment in a munge script.

reader 	 Character variable containing the name of the reader function that will be used to load the data. Contains a character(0) if no suitable reader was found.

* Note that some readers return more than one variable, usually with the listed variable name as prefix. This is true for for example the xls.reader and xlsx.reader.

Value

A data.frame listing the available data, with relevant meta data

See Also

load.project, show.project, project.config

Examples

library('ProjectTemplate')

## Not run: list.data()

[Package ProjectTemplate version 0.9.3 Index]