migrate.project {ProjectTemplate}	R Documentation

Migrates a project from a previous version of ProjectTemplate

Description

This function automatically performs all necessary steps to migrate an existing project so that it is compatible with this version of ProjectTemplate

Usage

migrate.project()

Value

No value is returned; this function is called for its side effects.

See Also

create.project

Examples

library('ProjectTemplate')

## Not run: migrate.project()

[Package ProjectTemplate version 0.9.3 Index]