migrate.template {ProjectTemplate}	R Documentation

Migrate a template to a new version of ProjectTemplate

Description

This function updates a skeleton project to the current version of ProjectTemplate.

Usage

migrate.template(template)

Arguments

template
Name of the template to upgrade.

[Package ProjectTemplate version 0.9.3 Index]