mp3.reader {ProjectTemplate}	R Documentation

Read an MP3 file with a .mp3 file extension.

Description

This function will load the specified MP3 file into memory using the tuneR package. This is useful for working with music files as a data set.

Usage

mp3.reader(data.file, filename, variable.name)

Arguments

data.file
The name of the data file to be read.

filename
The path to the data set to be loaded.

variable.name
The name to be assigned to in the global environment.

Value

No value is returned; this function is called for its side effects.

Examples

library('ProjectTemplate')

## Not run: mp3.reader('example.mp3', 'data/example.mp3', 'example')

[Package ProjectTemplate version 0.8.2 Index]