preinstalled.readers {ProjectTemplate}	R Documentation

Maps file types to the reader functions used to autoload them.

Description

This list stores a mapping from regular expressions that match file extensions for the file types supported by ProjectTemplate to the reader functions that implement autoloading for those formats. Any new file type must be appended to this dispatch table.

Usage

preinstalled.readers

Format

An object of class list of length 54.

[Package ProjectTemplate version 0.8.2 Index]