rdata.reader {ProjectTemplate}	R Documentation

Read an RData file with a .rdata or .rda file extension.

Description

This function will load the specified RData file into memory using the load function. This may generate many data sets simultaneously.

Usage

rdata.reader(data.file, filename, variable.name)

Arguments

data.file
The name of the data file to be read.

filename
The path to the data set to be loaded.

variable.name
The name to be assigned to in the global environment.

Value

No value is returned; this function is called for its side effects.

Examples

library('ProjectTemplate')

## Not run: rdata.reader('example.RData', 'data/example.RData', 'example')

[Package ProjectTemplate version 0.8.2 Index]