run.project {ProjectTemplate}	R Documentation

Run all of the analyses in the src directory.

Description

This function will run each of the analyses in the src directory in separate processes. At present, this is done serially, but future versions of this function will provide a means of running the analyses in parallel.

Usage

run.project()

Value

No value is returned; this function is called for its side effects.

Examples

library('ProjectTemplate')

## Not run: run.project()

[Package ProjectTemplate version 0.9.3 Index]