spss.reader {ProjectTemplate}	R Documentation

Read an SPSS file with a .sav file extension.

Description

This function will load the specified SPSS file into memory. It will convert the resulting list object into a data frame before inserting the data set into the global environment.

Usage

spss.reader(data.file, filename, variable.name)

Arguments

data.file
The name of the data file to be read.

filename
The path to the data set to be loaded.

variable.name
The name to be assigned to in the global environment.

Value

No value is returned; this function is called for its side effects.

Examples

library('ProjectTemplate')

## Not run: spss.reader('example.sav', 'data/example.sav', 'example')

[Package ProjectTemplate version 0.8.2 Index]