wsv.reader {ProjectTemplate}	R Documentation

Read a whitespace separated values (.wsv or .txt) file.

Description

This function will load a data set stored in the WSV file format into the specified global variable binding.

Usage

wsv.reader(data.file, filename, variable.name)

Arguments

data.file
The name of the data file to be read.

filename
The path to the data set to be loaded.

variable.name
The name to be assigned to in the global environment.

Value

No value is returned; this function is called for its side effects.

Examples

library('ProjectTemplate')

## Not run: wsv.reader('example.wsv', 'data/example.wsv', 'example')

[Package ProjectTemplate version 0.8.2 Index]