xlsx.reader {ProjectTemplate}	R Documentation

Read an Excel 2007 file with a .xlsx file extension.

Description

This function will load the specified Excel file into memory using the readxl package. Each sheet of the Excel workbook will be read into a separate variable in the global environment.

Usage

xlsx.reader(data.file, filename, workbook.name)

Arguments

data.file
The name of the data file to be read.

filename
The path to the data set to be loaded.

workbook.name
The name to be assigned to in the global environment.

Value

No value is returned; this function is called for its side effects.

Examples

library('ProjectTemplate')

## Not run: xlsx.reader('example.xlsx', 'data/example.xlsx', 'example')

[Package ProjectTemplate version 0.8.2 Index]